fix: make inline auth work in direct mode too (#520)
## Issue
Resolves
https://github.com/validator-labs/validator-plugin-aws/issues/517.

## Description
The original implementation of inline auth only worked in continuous
mode because the code that sets env vars for the AWS SDK was in
`Reconcile` and was therefore not invoked by `Validate` (which direct
mode invokes). This fixes it by moving that code to `Validate` so that
it will run during direct mode too.

Also makes the following minor changes:

* Renames `accessKeyPair` in the YAML to `credentials` to match other
plugins and leave room for other credential related data.
* Removes code that reads values for the `AWS_SESSION_TOKEN` env var
from secrets and inline config and sets the env var for this. This env
var isn't meant to be manually set. Instead, a role is meant to be
assumed. Our spec already allows STS to be enabled and a role ARN to be
configured when it's enabled.

---------

Signed-off-by: Matt Welke <matt.welke@spectrocloud.com>